First steps at making camera operations interruptable, by counting on gphoto to run the idle callback function and the cancel callback function.
Of course, no single gphoto driver today actually calls the idle function
to let qapp process the events, so we also try to process events in the cancel
callback function.

This all for now is single threaded, so we really need to rely on gphoto's cooperation for the GUI to remain responsive.

svn path=/trunk/kdegraphics/kamera/; revision=150079